LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-name-004-expected.txt


FAIL .target 1 assert_equals:
<div class="target" id="target1" data-expected-width="30"></div>
width expected 30 but got 0
FAIL .target 2 assert_equals:
<div class="target" id="target2" data-expected-width="30"></div>
width expected 30 but got 0
FAIL .target 3 assert_equals:
<div class="target" id="target3" data-expected-width="11"></div>
width expected 11 but got 0
